Separated mapper from rpl-collect example
