+++ apps/ids/Makefile.ids
@@ -1,1 +1,1 @@
-ids_src = ids-common.c
+ids_src = ids-common.c mapper.c
