+++ examples/ipv6/rpl-collect/network_mapper.c
@@ -1,0 +1,94 @@
+/*
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. Neither the name of the Institute nor the names of its contributors
+ *    may be used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE INSTITUTE AND CONTRIBUTORS ``AS IS'' AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED.  IN NO EVENT SHALL THE INSTITUTE OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
+ * OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+ * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+ * SUCH DAMAGE.
+ *
+ * This file is part of the Contiki operating system.
+ *
+ */
+
+#include "contiki.h"
+#include "contiki-lib.h"
+#include "contiki-net.h"
+#include "net/uip.h"
+#include "net/rpl/rpl.h"
+#include "net/rime/rimeaddr.h"
+
+#include "mapper.h"
+
+#include "net/netstack.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
+
+#define DEBUG DEBUG_PRINT
+#include "net/uip-debug.h"
+
+// Should not be nescesary, blaming strange build system
+void collect_common_set_sink(void) {}
+void collect_common_net_print(void) {}
+void collect_common_send(void) {}
+void collect_common_net_init(void) {}
+
+PROCESS(ids_server_example, "IDS Server router");
+AUTOSTART_PROCESSES(&ids_server_example, &mapper);
+
+/*---------------------------------------------------------------------------*/
+PROCESS_THREAD(ids_server_example, ev, data)
+{
+  uip_ipaddr_t ipaddr;
+  struct uip_ds6_addr *root_if;
+
+  PROCESS_BEGIN();
+
+  PROCESS_PAUSE();
+
+  PRINTF("IDS Server, compile time: %s\n", __TIME__);
+
+#if UIP_CONF_ROUTER
+  uip_ip6addr(&ipaddr, 0xaaaa, 0, 0, 0, 0, 0, 0, 1);
+
+  /* uip_ds6_set_addr_iid(&ipaddr, &uip_lladdr); */
+  uip_ds6_addr_add(&ipaddr, 0, ADDR_MANUAL);
+  root_if = uip_ds6_addr_lookup(&ipaddr);
+  if(root_if != NULL) {
+    rpl_dag_t *dag;
+    dag = rpl_set_root(RPL_DEFAULT_INSTANCE,(uip_ip6addr_t *)&ipaddr);
+    uip_ip6addr(&ipaddr, 0xaaaa, 0, 0, 0, 0, 0, 0, 0);
+    rpl_set_prefix(dag, &ipaddr, 64);
+    PRINTF("created a new RPL dag\n");
+  } else {
+    PRINTF("failed to create a new RPL DAG\n");
+  }
+#else
+#error "No router"
+#endif /* UIP_CONF_ROUTER */
+
+  /* The data sink runs with a 100% duty cycle in order to ensure high
+     packet reception rates. */
+  NETSTACK_RDC.off(1);
+
+  PROCESS_END();
+}
+/*---------------------------------------------------------------------------*/
